+++ src/WhatsAppClient.ts
@@ -1,4 +1,5 @@
 import axios from 'axios'
+import FormData from 'form-data'
 import { WhatsAppConfig, WhatsAppResultContract } from '@ioc:Adonis/Addons/WhatsApp'
 
 type WhatsAppResult = {
@@ -25,16 +26,48 @@ export default class WhatsAppClient {
     recipient_type: 'individual',
   }
 
-  public async send(method: 'GET' | 'POST' | 'PUT', data: Record<string, any>, parse = true) {
+  public async send(data: Record<string, any>, parse = true) {
+    const { timeout, phoneNumberId, graphUrl, graphVersion } = this.config
+
     const response = await axios({
-      method,
-      url: this.config.graphUrl + '/' + this.config.graphVersion + '/' + this.config.phoneNumberId,
-      timeout: this.config.timeout,
+      method: 'POST',
+      url: `${graphUrl}/${graphVersion}/${phoneNumberId}/messages`,
+      timeout,
       headers: this.headers,
       data: { ...this.skeleton, ...data },
+      responseType: 'json',
     })
 
     return parse ? WhatsAppClient.parse(response.data) : response.data
+  }
+
+  public async media(media: string) {
+    const { timeout, graphUrl, graphVersion } = this.config
+
+    const response = await axios({
+      method: 'GET',
+      url: `${graphUrl}/${graphVersion}/${media}`,
+      timeout,
+      headers: this.headers,
+      responseType: 'json',
+    })
+
+    return response.data
+  }
+
+  public async upload(form: FormData) {
+    const { timeout, phoneNumberId, graphUrl, graphVersion } = this.config
+
+    const response = await axios({
+      method: 'POST',
+      url: `${graphUrl}/${graphVersion}/${phoneNumberId}/media`,
+      timeout,
+      headers: { ...form.getHeaders(), ...this.headers },
+      data: form,
+      responseType: 'json',
+    })
+
+    return response.data
   }
 
   private static parse(data: WhatsAppResult): WhatsAppResultContract {