+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## 0.2.0
+
+- Added method for media:
+  - `WhatsApp.uploadMedia(source)`
+  - `WhatsApp.downloadMedia(media, [options])`
+- Options in `sendText` method are optional
+- Type in contact options fixed
+- Added return type in `translateType` helper
+- Send message endpoint fixed
+- Increase default timeout configuration
+- `sendSticker` method fixed
+
+## 0.1.1
+
+Added preload file
+
 ## 0.1.0
 
 First version